refactor: move ValidateCapabilities from v1beta1helper to gardenerutils
