+++ pkg/provider-local/apis/local/v1alpha1/types_cloudprofile.go
@@ -48,8 +48,3 @@ type MachineImageFlavor struct {
 	// Capabilities that are supported by the identifier in this set.
 	Capabilities gardencorev1beta1.Capabilities `json:"capabilities,omitempty"`
 }
-
-// GetCapabilities returns the Capabilities of a MachineImageFlavor
-func (cs MachineImageFlavor) GetCapabilities() gardencorev1beta1.Capabilities {
-	return cs.Capabilities
-}
